refactor: remove unnecesary classes and attributes
Date-related attributes were removed from Entry and Patient classes
in favor of the inherited CreatedAt attribute (Parent Class: BaseEntity).
Both IdentificationType and LinkRequestStatus classes were deleted to be replaced
for ENUMs in a future commit.

An "Active" attribute of type bool was added to Patient and Therapist classes
to represent the deletion of either one of said entities.

"HabitSchedule" class was created to replace the Days attribute of type string
in the PatientHabit class.